+++ lighten.sh
@@ -8,7 +8,7 @@ export CONSOLE_THEME="light"
 echo "$this_file_parent_dir/lighten.sh" > ~/.config/colorshift/target_file.txt
 
 # modify gnome-terminal default profile
-#"$this_file_parent_dir/set_gnome_default" Light
+"$this_file_parent_dir/set_gnome_default" Light
 
 # apply dircolors
 source "$this_file_parent_dir/dircolors.sh"
@@ -18,8 +18,11 @@ dircolors_file ~/.config/dircolors/dircolors-solarized/dircolors.ansi-dark
 ZSH_AUTOSUGGEST_HIGHLIGHT_STYLE='fg=180'
 
 # zsh easy motion
-EASY_MOTION_DIM="fg=60,bold"
+EASY_MOTION_DIM="fg=180,bold"
 
 # urxvt
-cp ~/.config/xresources/Xresources.light ~/.Xresources
-xrdb -load ~/.Xresources
+if [ -f ~/.config/xresources/Xresources.light]
+then
+    cp ~/.config/xresources/Xresources.light ~/.Xresources
+    xrdb -load ~/.Xresources
+fi
